refactor: collapse duplicate dispute entrypoints to one safe implementation
