+++ contracts/escrow/src/lib.rs
@@ -1407,6 +1407,8 @@ impl Escrow {
     /// - Only the assigned arbiter can resolve disputes
     /// - Split amounts must exactly match available balance
     /// - Updates released_amount and refunded_amount atomically
+    /// - Verifies released and refunded accounting consumes the funded balance
+    /// - Adds one pending reputation credit for the freelancer on completion
     /// - Emits dispute resolution event for indexers
     /// - Sets final contract status based on resolution outcome
     pub fn resolve_dispute(
@@ -1444,11 +1446,26 @@ impl Escrow {
                 .unwrap_or_else(|e| env.panic_with_error(e));
 
         // Update contract accounting
-        contract.refunded_amount += client_payout;
-        contract.released_amount += freelancer_payout;
+        contract.refunded_amount = safe_add_amounts(contract.refunded_amount, client_payout)
+            .unwrap_or_else(|| env.panic_with_error(EscrowError::PotentialOverflow));
+        contract.released_amount = safe_add_amounts(contract.released_amount, freelancer_payout)
+            .unwrap_or_else(|| env.panic_with_error(EscrowError::PotentialOverflow));
+
+        if safe_add_amounts(contract.released_amount, contract.refunded_amount)
+            != Some(contract.funded_amount)
+        {
+            env.panic_with_error(EscrowError::AccountingInvariantViolated);
+        }
 
         // Set final status
         contract.status = dispute::final_status_after_resolution(&contract);
+        if contract.status == ContractStatus::Completed {
+            let pending_key = DataKey::PendingReputationCredits(contract.freelancer.clone());
+            let pending: i128 = env.storage().persistent().get(&pending_key).unwrap_or(0);
+            let pending = safe_add_amounts(pending, 1)
+                .unwrap_or_else(|| env.panic_with_error(EscrowError::PotentialOverflow));
+            env.storage().persistent().set(&pending_key, &pending);
+        }
 
         env.storage()
             .persistent()
